requirements.txt
flask
flask-script
flask-migrate
flask-restful
flask-marshmallow
Flask-JWT
marshmallow-sqlalchemy
flask-bcrypt
